[BREAKING] Major refactor (see notes)
Rewrote rule engine to be simpler and more flexible. Now you can just provide a "Rule Template" which creates a ruleset containing a list of rules. Each RuleSet gets the full configuration object and can do whatever it wants as long as it returns an array of rules.

Responsiveness is improved: Any rule can be made responsive by just passing in an array of styles that you wish to make responsive.

Disabling rules: Now just pass in an array of ruleset "Keys" to disable the given ruleset.

Simpler CLI implementation.

Add example project.

Lots of other cleanup and improvements and tests.